# Add tests to verify performance of SmartFields