Testing the performance of the `SmartDeviceField<T,ACCESS>` vs `skt::mesh::NgpField<T>` on device.  Results below from two instances below using a Cuda11.2 Release build:

```console
[ RUN      ] TestSmartField.device_performance_smart_field
[       OK ] TestSmartField.device_performance_smart_field (396 ms)
[ RUN      ] TestSmartField.device_performance_ngp_field
[       OK ] TestSmartField.device_performance_ngp_field (373 ms)
````

``` console
[ RUN      ] TestSmartField.device_performance_smart_field
[       OK ] TestSmartField.device_performance_smart_field (387 ms)
[ RUN      ] TestSmartField.device_performance_ngp_field
[       OK ] TestSmartField.device_performance_ngp_field (423 ms)
```